+++ casia/src/views/MenuViews/ArchivesView.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import MainTemplate from '../../components/templates/MainTemplate/MainTemplate';
+
+const ArchivesView = () => {
+  return (
+    <MainTemplate>
+      <h1> Hello There ArchivesView!</h1>
+    </MainTemplate>
+  );
+};
+
+export default ArchivesView;